move utility funcs into utilities.py
