+++ hyperpackage/hyperpackage/hyperpack_creation.py
@@ -1,9 +1,12 @@
-import json
 import os
-import shutil
-import yaml
 from datetime import datetime
 from hyperpackage.flavor.pytorch import torch_onnx_export
+from hyperpackage.utilities import (
+    generate_folder_name,
+    write_json,
+    write_yaml,
+    zip_study,
+)
 
 SUPPORTED_MODEL_FLAVORS = ["automl"]
 
@@ -58,32 +61,3 @@ def make_hyperpack_path(base_dir: str, name: str) -> str:
     hyperpack_folder_name = name + ".hyperpack"
     path = os.path.join(base_dir, hyperpack_folder_name)
     return path
-
-
-def write_json(dictionary, json_file_path):
-    with open(json_file_path, "w") as json_file:
-        json_file.write(json.dumps(dictionary))
-
-
-def write_yaml(dictionary: dict, yaml_file_path: str):
-    with open(yaml_file_path, "w") as stream:
-        yaml.dump(dictionary, stream)
-
-
-def zip_study(folder_path):
-    shutil.make_archive(folder_path, "zip", folder_path)
-
-
-def generate_folder_name(
-    trial_id: int = 0,
-    name: str = None,
-    format_precision: str = "06",
-    suffix: str = "trial",
-):
-    prefix = format(trial_id, format_precision)
-    if name is not None:
-        folder_name = f"{prefix}-{name}-{suffix}"
-    else:
-        folder_name = f"{prefix}-{suffix}"
-
-    return folder_name
